win,unix: uv_buf_init() accepts size_t
Change the signature of uv_buf_init() from (char*, unsigned int) to
(char*, size_t). This is to accommodate the fact that the "len" property
of uv_buf_t is a size_t, and fixes a problem where "len" is truncated
for large "base".

Fixes: https://github.com/libuv/libuv/issues/1233
PR-URL: https://github.com/libuv/libuv/pull/1236

+++ include/uv.h
@@ -432,7 +432,7 @@ UV_EXTERN int uv_recv_buffer_size(uv_handle_t* handle, int* value);
 
 UV_EXTERN int uv_fileno(const uv_handle_t* handle, uv_os_fd_t* fd);
 
-UV_EXTERN uv_buf_t uv_buf_init(char* base, unsigned int len);
+UV_EXTERN uv_buf_t uv_buf_init(char* base, size_t len);
 
 
 #define UV_STREAM_FIELDS                                                      \
