+++ src/uv-common.c
@@ -137,7 +137,7 @@ size_t uv_loop_size(void) {
 }
 
 
-uv_buf_t uv_buf_init(char* base, unsigned int len) {
+uv_buf_t uv_buf_init(char* base, size_t len) {
   uv_buf_t buf;
   buf.base = base;
   buf.len = len;
